feat(FN-1536): refactor tfm-api route validator logic
