+++ trade-finance-manager-api/src/v1/validation/route-validators/update-report-status-payload-validation/index.test.ts
@@ -1,0 +1,215 @@
+import { Request } from 'express';
+import { validationResult } from 'express-validator';
+import { createRequest, createResponse } from 'node-mocks-http';
+import { ObjectId } from 'mongodb';
+import { updateReportStatusPayloadValidation } from '.';
+import { UTILISATION_REPORT_RECONCILIATION_STATUS } from '../../../../constants';
+import { TfmSessionUser } from '../../../../types/tfm-session-user';
+import { ReportIdentifier, ReportWithStatus, UtilisationReportReconciliationStatus } from '../../../../types/utilisation-reports';
+
+type ValidPayloadBodyOpts = {
+  user?: TfmSessionUser;
+  reportsWithStatus?: ReportWithStatus[];
+  status?: UtilisationReportReconciliationStatus;
+  report?: ReportIdentifier;
+};
+
+describe('updateReportStatusPayloadValidation', () => {
+  const getValidPayloadBody = (opts: ValidPayloadBodyOpts) => ({
+    user: opts.user ?? {},
+    reportsWithStatus: opts.reportsWithStatus ?? [
+      {
+        status: opts.status ?? UTILISATION_REPORT_RECONCILIATION_STATUS.PENDING_RECONCILIATION,
+        report: opts.report ?? {
+          month: 5,
+          year: 2023,
+          bankId: '123',
+        },
+      },
+    ],
+  });
+
+  const getUpdateReportStatusPayloadValidationResult = async (req: Request) => {
+    const res = createResponse();
+    const next = jest.fn();
+
+    const validators = updateReportStatusPayloadValidation.map((validator) => {
+      if (Array.isArray(validator)) {
+        return validator.flatMap((subValidator) => subValidator(req, res, next));
+      }
+      return validator(req, res, next);
+    });
+    await Promise.all(validators);
+
+    return validationResult(req).array();
+  };
+
+  describe('for a payload with a bank id, month and year combination', () => {
+    const validPayload = {
+      month: 5,
+      year: 2023,
+      bankId: '123',
+    };
+
+    it.each([
+      // @ts-expect-error month is supposed to be the incorrect type for this test
+      ['month is a string', getValidPayloadBody({ report: { ...validPayload, month: '5' } })],
+      ['month is less than 1', getValidPayloadBody({ report: { ...validPayload, month: 0 } })],
+      ['month is greater than 12', getValidPayloadBody({ report: { ...validPayload, month: 13 } })],
+      // @ts-expect-error year is supposed to be the incorrect type for this test
+      ['year is a string', getValidPayloadBody({ report: { ...validPayload, year: '2023' } })],
+      ['year is less than 0', getValidPayloadBody({ report: { ...validPayload, year: -1 } })],
+      // @ts-expect-error bankId is supposed to be the incorrect type for this test
+      ['bankId is a number', getValidPayloadBody({ report: { ...validPayload, bankId: 123 } })],
+    ])('returns a single error when one of the parameters in invalid (%s)', async (_, payload) => {
+      // Arrange
+      const req = createRequest({ body: payload });
+
+      // Act
+      const errors = await getUpdateReportStatusPayloadValidationResult(req);
+
+      // Assert
+      expect(errors.length).toEqual(1);
+    });
+
+    it('returns no errors when the payload is valid', async () => {
+      // Arrange
+      const body = getValidPayloadBody({ report: validPayload });
+      const req = createRequest({ body });
+
+      // Act
+      const errors = await getUpdateReportStatusPayloadValidationResult(req);
+
+      // Assert
+      expect(errors.length).toEqual(0);
+    });
+  });
+
+  describe('for a payload with a report id', () => {
+    const validMongoId = '5ce819935e539c343f141ece';
+
+    it('returns a single error when the report id is not a valid mongo id', async () => {
+      // Arrange
+      const report = { id: '123' };
+      const body = getValidPayloadBody({ report });
+      const req = createRequest({ body });
+
+      // Act
+      const errors = await getUpdateReportStatusPayloadValidationResult(req);
+
+      // Assert
+      expect(errors.length).toEqual(1);
+    });
+
+    it('returns a single error when the report id is not a string', async () => {
+      // Arrange
+      const report = { id: new ObjectId(validMongoId) };
+      // @ts-expect-error report is supposed to be invalid for this test
+      const body = getValidPayloadBody({ report });
+      const req = createRequest({ body });
+
+      // Act
+      const errors = await getUpdateReportStatusPayloadValidationResult(req);
+
+      // Assert
+      expect(errors.length).toEqual(1);
+    });
+
+    it('returns no errors when the payload is valid', async () => {
+      // Arrange
+      const report = { id: validMongoId };
+      const body = getValidPayloadBody({ report });
+      const req = createRequest({ body });
+
+      // Act
+      const errors = await getUpdateReportStatusPayloadValidationResult(req);
+
+      // Assert
+      expect(errors.length).toEqual(0);
+    });
+  });
+
+  it('returns no errors for a payload which uses a combination of report identifiers', async () => {
+    // Arrange
+    const validMongoId = '5ce819935e539c343f141ece';
+    const reportWithId = { id: validMongoId };
+    const reportWithBankId = {
+      month: 1,
+      year: 2023,
+      bankId: '123',
+    };
+    const reportsWithStatus = [
+      {
+        status: UTILISATION_REPORT_RECONCILIATION_STATUS.RECONCILIATION_COMPLETED,
+        report: reportWithId,
+      },
+      {
+        status: UTILISATION_REPORT_RECONCILIATION_STATUS.RECONCILIATION_COMPLETED,
+        report: reportWithBankId,
+      },
+    ];
+    const body = getValidPayloadBody({ reportsWithStatus });
+    const req = createRequest({ body });
+
+    // Act
+    const errors = await getUpdateReportStatusPayloadValidationResult(req);
+
+    // Assert
+    expect(errors.length).toEqual(0);
+  });
+
+  it('returns a single error when the user is not an object', async () => {
+    // Arrange
+    const user = 'Test user';
+    // @ts-expect-error user is supposed to be invalid for this test
+    const body = getValidPayloadBody({ user });
+    const req = createRequest({ body });
+
+    // Act
+    const errors = await getUpdateReportStatusPayloadValidationResult(req);
+
+    // Assert
+    expect(errors.length).toEqual(1);
+  });
+
+  it('returns a single error when reportsWithStatus is an empty array', async () => {
+    // Arrange
+    const reportsWithStatus: ReportWithStatus[] = [];
+    const body = getValidPayloadBody({ reportsWithStatus });
+    const req = createRequest({ body });
+
+    // Act
+    const errors = await getUpdateReportStatusPayloadValidationResult(req);
+
+    // Assert
+    expect(errors.length).toEqual(1);
+  });
+
+  it('returns a single error when the status is not a valid status', async () => {
+    // Arrange
+    const status = 'INVALID_STATUS' as UtilisationReportReconciliationStatus;
+    const body = getValidPayloadBody({ status });
+    const req = createRequest({ body });
+
+    // Act
+    const errors = await getUpdateReportStatusPayloadValidationResult(req);
+
+    // Assert
+    expect(errors.length).toEqual(1);
+  });
+
+  it.each([
+    { status: UTILISATION_REPORT_RECONCILIATION_STATUS.RECONCILIATION_COMPLETED },
+    { status: UTILISATION_REPORT_RECONCILIATION_STATUS.PENDING_RECONCILIATION },
+  ])('returns no errors when the status is $status', async ({ status }) => {
+    // Arrange
+    const body = getValidPayloadBody({ status });
+    const req = createRequest({ body });
+
+    // Act
+    const errors = await getUpdateReportStatusPayloadValidationResult(req);
+
+    // Assert
+    expect(errors.length).toEqual(0);
+  });
+});
